refactor: separate labelstudio class to several
