+++ convert/convert/converters/labelstudio/labelstudio_format.py
@@ -1,25 +1,16 @@
-import tempfile
 from pathlib import Path
-from typing import Any, Dict, List, NamedTuple, Optional, Tuple
+from typing import Any, Dict, List, Optional
 
 import requests
-from botocore.client import BaseClient
 from fastapi import HTTPException, status
-from tenant_dependency import TenantData
 
 from convert.config import settings
-from convert.converters.base_format.models import annotation_practic
 from convert.converters.base_format.models.annotation import (
     BadgerdocAnnotation,
 )
 from convert.converters.base_format.models.manifest import Manifest
 from convert.converters.base_format.models.tokens import BadgerdocToken, Page
-from convert.converters.labelstudio.annotation_converter_practic import (
-    AnnotationConverterToTheory,
-)
 from convert.logger import get_logger
-from convert.models.common import S3Path
-from convert.converters.labelstudio.utils import combine
 
 from .models.annotation import (
     Annotation,
@@ -34,12 +25,6 @@ from .models.annotation import (
 
 LOGGER = get_logger(__file__)
 LOGGER.setLevel("DEBUG")
-
-
-class BadgerdocData(NamedTuple):
-    page: Page
-    annotation: BadgerdocAnnotation
-    manifest: Manifest
 
 
 class LabelStudioFormat:
@@ -57,7 +42,7 @@ class LabelStudioFormat:
     def from_badgerdoc(
         self,
         badgerdoc_tokens: Page,
-        badgerdoc_annotations: BadgerdocAnnotation,
+        badgerdoc_annotations: Optional[BadgerdocAnnotation],
         badgerdoc_manifest: Optional[Manifest],
         request_headers: Dict[str, str],
     ):
@@ -65,10 +50,34 @@ class LabelStudioFormat:
             [self.form_token_text(obj) for obj in badgerdoc_tokens.objs]
         )
 
-        objs = self.convert_annotation_from_bd(
-            badgerdoc_annotations, badgerdoc_tokens.objs, text
+        annotation = None
+        if badgerdoc_annotations:
+            objs = self.convert_annotation_from_bd(
+                badgerdoc_annotations, badgerdoc_tokens.objs, text
+            )
+            relations = self.convert_relation_from_bd(badgerdoc_annotations)
+            annotation = Annotation(
+                id=self.DEFAULT_ID_FOR_ONE_ANNOTATION, result=objs + relations
+            )
+
+        meta = Meta()
+        if badgerdoc_manifest:
+            meta = self.form_meta(badgerdoc_manifest, request_headers)
+
+        item = ModelItem(
+            annotations=[annotation],
+            predictions=[],
+            data=Data(text=text),
+            meta=meta,
         )
-        relations = self.convert_relation_from_bd(badgerdoc_annotations)
+
+        self.labelstudio_data.__root__.append(item)
+
+    def form_meta(
+            self,
+            badgerdoc_manifest: Optional[Manifest],
+            request_headers: Dict[str, str],
+        ) -> Meta:
         document_links = (
             self.convert_document_links_from_bd(badgerdoc_manifest)
             if badgerdoc_manifest
@@ -96,22 +105,11 @@ class LabelStudioFormat:
             categories_linked_with_taxonomies=categories_linked_with_taxonomies,  # noqa
             request_headers=request_headers,
         )
-        annotation = Annotation(
-            id=self.DEFAULT_ID_FOR_ONE_ANNOTATION, result=objs + relations
-        )
-
-        item = ModelItem(
-            annotations=[annotation],
-            predictions=[],
-            data=Data(text=text),
-            meta=Meta(
-                labels=document_labels,
-                relations=document_links,
-                categories_to_taxonomy_mapping=categories_to_taxonomy_mapping,
-            ),
-        )
-
-        self.labelstudio_data.__root__.append(item)
+        return Meta(
+            labels=document_labels,
+            relations=document_links,
+            categories_to_taxonomy_mapping=categories_to_taxonomy_mapping,
+        ) 
 
     @classmethod
     def convert_annotation_from_bd(
@@ -340,156 +338,3 @@ class LabelStudioFormat:
             result[taxonomy_id_of_taxon].append(taxon_obj["id"])
 
         return result
-
-
-class WipBadgerdocToLabelstudioConverter:
-    def __init__(
-        self,
-        s3_client: BaseClient,
-        current_tenant: str,
-        token_data: TenantData,
-    ) -> None:
-        self.s3_client = s3_client
-        self.current_tenant = current_tenant
-        self.token_data = token_data
-        self.request_headers = {
-            "X-Current-Tenant": self.current_tenant,
-            "Authorization": f"Bearer {self.token_data.token}",
-        }
-
-    def execute(
-        self,
-        s3_input_tokens: S3Path,
-        s3_input_manifest: S3Path,
-        s3_output_annotation: S3Path,
-    ) -> None:
-        downloader = BadgerdocDownloader(self.s3_client, s3_input_tokens, s3_input_manifest)
-        (
-            pages,
-            annotations,
-            manifest,
-        ) = downloader.download()
-        labelstudio_pages = self.convert_to_labelestudio(pages, annotations, manifest)
-        labelstudio_combined_pages_data = combine(labelstudio_pages)
-        labelstudio_combined = LabelStudioFormat()
-        labelstudio_combined.labelstudio_data = labelstudio_combined_pages_data
-        uploader = LabelStudioUploader(self.s3_client)
-        uploader.upload(labelstudio_combined, s3_output_annotation)
-
-    def convert_to_labelestudio(
-            self,
-            pages: List[Page],
-            annotations: Dict[int, BadgerdocAnnotation],
-            manifest: Manifest
-        ) -> List[LabelStudioModel]:
-        labelstudio_pages = []
-        for page in pages:
-            labelstudio = LabelStudioFormat()
-            labelstudio.from_badgerdoc(
-                page,
-                annotations[page.page_num],
-                manifest,
-                self.request_headers,
-            )
-            labelstudio_pages.append(labelstudio.labelstudio_data)
-        return labelstudio_pages
-
-
-class BadgerdocDownloader:
-    def __init__(self, s3_client: BaseClient, s3_input_tokens: S3Path, s3_input_manifest: S3Path):
-        self.s3_input_tokens = s3_input_tokens
-        self.s3_input_manifest = s3_input_manifest
-        self.s3_client = s3_client
-
-    def download(
-        self,
-    ) -> Tuple[List[Page], Dict[int, BadgerdocAnnotation], Manifest]:
-        with tempfile.TemporaryDirectory() as tmp_dirname:
-            tmp_dir = Path(tmp_dirname)
-
-            pages = self.get_token_pages(tmp_dir)
-            manifest = self.get_manifest(tmp_dir)
-            annotations = self.get_annotations(manifest, tmp_dir)
-
-            return pages, annotations, manifest
-
-    def get_token_pages(self, tmp_dir) -> List[Page]:
-        token_page_files = self.download_all_token_pages(self.s3_input_tokens, tmp_dir)
-        pages = []
-        for token_page_file in token_page_files:
-            pages.append(Page.parse_file(token_page_file))
-        return pages
-
-    def get_manifest(self, tmp_dir) -> Manifest:
-        manifest_file = self.download_file_from_s3(
-            self.s3_input_manifest, tmp_dir
-        )
-        return Manifest.parse_file(tmp_dir / manifest_file.name)
-
-    def get_annotations(self, manifest: Manifest, tmp_dir) -> Dict[int, BadgerdocAnnotation]:
-        annotation_files = self.download_annotations(
-            manifest_s3_path=self.s3_input_manifest,
-            manifest=manifest,
-            tmp_dir=tmp_dir
-        )
-        annotations = {}
-        for page_num, annotation_file in annotation_files.items():
-            # annotations[int(page_num)] = annotation_practic.BadgerdocAnnotation.parse_file(annotation_file)
-            annotations[int(page_num)] = AnnotationConverterToTheory(
-                practic_annotations=annotation_practic.BadgerdocAnnotation.parse_file(  # noqa
-                    annotation_file
-                )
-            ).convert()
-        return annotations
-
-    def download_all_token_pages(self, s3_path: S3Path, tmp_dir: Path) -> List[Path]:
-        response = self.s3_client.list_objects(Bucket=s3_path.bucket, Prefix=s3_path.path)
-        pages = (obj['Key'] for obj in response['Contents'])
-        page_files = []
-        for page in pages:
-            page_path = S3Path(bucket=s3_path.bucket, path=f"{page}")
-            page_files.append(self.download_file_from_s3(s3_path=page_path, tmp_dir=tmp_dir))
-        return page_files
-
-    def download_file_from_s3(self, s3_path: S3Path, tmp_dir: Path) -> Path:
-        local_file_path = tmp_dir / Path(s3_path.path).name
-        self.s3_client.download_file(
-            s3_path.bucket,
-            s3_path.path,
-            str(local_file_path),
-        )
-        return local_file_path
-
-    def download_annotations(self, manifest_s3_path: S3Path, manifest: Manifest, tmp_dir: Path) -> Dict[str, S3Path]:
-        pages = {}
-        for page_num, page_file in manifest.pages.items():
-            page_s3_path = self.form_absolute_path_for_annotation(manifest_s3_path, page_file)
-            pages[page_num] = self.download_file_from_s3(page_s3_path, tmp_dir)
-        return pages
-
-    def form_absolute_path_for_annotation(self, manifest_s3_path: S3Path, page_file: str) -> S3Path:
-        absolute_path = f"{Path(manifest_s3_path.path).parent}/{page_file}.json"
-        return S3Path(bucket=manifest_s3_path.bucket, path=absolute_path)
-
-
-class LabelStudioUploader:
-    LABELSTUDIO_FILENAME = "ls_format.json"
-
-    def __init__(self, s3_client: BaseClient):
-        self.s3_client = s3_client
-
-    def upload(
-        self,
-        labelstudio: LabelStudioFormat,
-        s3_output_annotation: S3Path,
-    ) -> None:
-        with tempfile.TemporaryDirectory() as tmp_dirname:
-            badgerdoc_annotations_path = Path(tmp_dirname) / Path(
-                self.LABELSTUDIO_FILENAME
-            )
-            labelstudio.export_json(badgerdoc_annotations_path)
-            self.s3_client.upload_file(
-                str(badgerdoc_annotations_path),
-                s3_output_annotation.bucket,
-                s3_output_annotation.path,
-            )
