+++ convert/convert/models/labelstudio.py
@@ -21,12 +21,5 @@ class LabelStudioRequest(BaseModel):
 
 class BadgerdocToLabelStudioRequest(BaseModel):
     input_tokens: S3Path
-    input_annotation: S3Path
-    input_manifest: S3Path
-    output_annotation: S3Path
-
-
-class WipBadgerdocToLabelStudioRequest(BaseModel):
-    input_tokens: S3Path
     input_manifest: S3Path
     output_annotation: S3Path
