+++ convert/convert/routers/labelstudio.py
@@ -7,14 +7,13 @@ from convert.converters.labelstudio.badgerdoc_to_labelstudio_converter import (
 from convert.converters.labelstudio.labelstudio_to_badgerdoc_converter import (
     LabelstudioToBadgerdocConverter,
 )
-from convert.converters.labelstudio.wip_badgerdoc_to_labelstudio_converter import (
-    WipBadgerdocToLabelstudioConverter,
+from convert.converters.labelstudio.badgerdoc_to_labelstudio_converter import (
+    BadgerdocToLabelstudioConverter,
 )
 
 from convert.models.labelstudio import (
-    BadgerdocToLabelStudioRequest,
     LabelStudioRequest,
-    WipBadgerdocToLabelStudioRequest,
+    BadgerdocToLabelStudioRequest,
 )
 from fastapi import APIRouter, Depends, Header, status
 from tenant_dependency import TenantData, get_tenant_info
@@ -55,32 +54,10 @@ def import_labelstudio(
 )
 def export_labelstudio(
     request: BadgerdocToLabelStudioRequest,
-    current_tenant: Optional[str] = Header(None, alias="X-Current-Tenant"),
-    token_data: TenantData = Depends(tenant),
-) -> None:
-    bd_to_labelstudio_use_case = BadgerdocToLabelstudioConverter(
-        s3_client=minio_client,
-        current_tenant=current_tenant,
-        token_data=token_data,
-    )
-    bd_to_labelstudio_use_case.execute(
-        s3_input_tokens=request.input_tokens,
-        s3_input_annotations=request.input_annotation,
-        s3_input_manifest=request.input_manifest,
-        s3_output_annotation=request.output_annotation,
-    )
-
-
-@router.post(
-    "/wip_export",
-    status_code=status.HTTP_201_CREATED,
-)
-def export_labelstudio(
-    request: WipBadgerdocToLabelStudioRequest,
     current_tenant: str = Header(None, alias="X-Current-Tenant"),
     token_data: TenantData = Depends(tenant),
 ) -> None:
-    bd_to_labelstudio_converter = WipBadgerdocToLabelstudioConverter(
+    bd_to_labelstudio_converter = BadgerdocToLabelstudioConverter(
         s3_client=minio_client,
         current_tenant=current_tenant,
         token_data=token_data
